cs export

+++ app/Http/Controllers/CustomerSatisfactionController.php
@@ -4,15 +4,18 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Facades\CustomLog;
+use Illuminate\Support\Str;
 use App\Models\CustomerSatisfaction;
 use App\Models\SatisfactionColumn;
 use Illuminate\Support\Facades\Auth;
+use App\Exports\CustomerSatisfactionExport;
+use Maatwebsite\Excel\Facades\Excel;
 
 class CustomerSatisfactionController extends Controller
 {
     public function index()
     {
-        if(empty(session('standard'))){
+        if(session('standard') == null || session('standard_name') != "9001"){
             return redirect('/');
         }
 
@@ -27,12 +30,19 @@ class CustomerSatisfactionController extends Controller
 
     public function create()
     {
+        if(session('standard') == null || session('standard_name') != "9001"){
+            return redirect('/');
+        }
+
+        $this->authorize('create', CustomerSatisfaction::class);
+
         $poll = SatisfactionColumn::where('team_id', Auth::user()->current_team_id)->whereNotNull('name')->get();
         return view('system_processes.customer_satisfaction.create', compact('poll'));
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', CustomerSatisfaction::class);
 
         $cs = $request->except('_token');
         $cs['standard_id'] = session('standard');
@@ -53,20 +63,29 @@ class CustomerSatisfactionController extends Controller
         $poll = SatisfactionColumn::where('team_id', Auth::user()->current_team_id)->whereNotNull('name')->get();
 
         $cs->columns = $poll;
-        $cs->average = $cs->average();
+        $cs->average = round($cs->average(), 1);
         return $cs;
     }
 
     public function edit($id)
     {
-        $poll = SatisfactionColumn::where('team_id', Auth::user()->current_team_id)->whereNotNull('name')->get();
+        if(session('standard') == null || session('standard_name') != "9001"){
+            return redirect('/');
+        }
+
         $cs = CustomerSatisfaction::findOrFail($id);
+        $poll = SatisfactionColumn::where('team_id', Auth::user()->current_team_id)->whereNotNull('name')->get();
+
+        $this->authorize('update', $cs);
+        
         return view('system_processes.customer_satisfaction.edit', compact('poll', 'cs'));
     }
 
     public function update(Request $request, $id)
     {
         $cs = CustomerSatisfaction::findOrFail($id);
+
+        $this->authorize('update', $cs);
         
         try{
             $cs->update($request->except('_token', '_method'));
@@ -83,6 +102,8 @@ class CustomerSatisfactionController extends Controller
     {
         $cs = CustomerSatisfaction::findOrFail($id);
 
+        $this->authorize('delete', $cs);
+
         try{
             CustomerSatisfaction::destroy($id);
             CustomLog::info('Anketa za klijenta "'.$cs->customer.'" uklonjena, '.Auth::user()->name.', '.Auth::user()->username.', '.date('d.m.Y H:i:s'), Auth::user()->currentTeam->name);
@@ -91,5 +112,13 @@ class CustomerSatisfactionController extends Controller
             CustomLog::warning('Neuspeli pokušaj brisanja ankete za korisnika '.$cs->customer.', '.Auth::user()->name.', '.Auth::user()->username.', '.date('d.m.Y H:i:s').', Greška: '.$e->getMessage(), Auth::user()->currentTeam->name);
             return back()->with('status', array('danger', __('Došlo je do greške! Pokušajte ponovo.')));
         }
+    }
+
+    public function export()
+    {
+        if(empty(session('standard'))){
+            return redirect('/');
+        }
+        return Excel::download(new CustomerSatisfactionExport, Str::snake(__('Zadovoljstvo korisnika')).'_'.session('standard_name').'.xlsx');
     }
 }
